### 2019-09-26 #1767 @whedon
[PRE REVIEW]: AliasFinder: A Python script to search for the true planetary frequency within radial velocity data
**Submitting author:** @StephanStock (<a href="http://orcid.org/0000-0002-1166-9338">Stephan Stock</a>)
**Repository:** <a href="https://github.com/JonasKemmer/AliasFinder" target ="_blank">https://github.com/JonasKemmer/AliasFinder</a>
**Version:** v1.0.0
