### 2019-09-26 #1767 @arfon
@StephanStock - this looks to be a duplicate of https://github.com/openjournals/joss-reviews/issues/1724 which was rejected as being out of scope. Could you clarify why there has been a resubmission?

/ cc @danielskatz
