### 2019-09-26 #1767 @StephanStock
Hello,
as requested in #1724 we implemented the pip install for our software which was the reason why it was rejected. It is now  'designed for maintainable extension'.

/cc @JonasKemmer @xuanxu
