### 2019-09-27 #1767 @xuanxu
I think in its current form it qualifies as a maintainable research software, so I'm moving this forward
