### 2019-09-27 #1767 @xuanxu
:wave: @ygrange - would you be willing to review this submission for JOSS? I think your knowledge of Python and Astronomy makes you a good fit for this submission.
