### 2019-09-27 #1767 @xuanxu
:wave: @wtgee - would you be willing to review this submission for JOSS? I think your knowledge of Python and Exoplanets makes you a great fit for this submission.
