### 2019-09-28 #1767 @wtgee
@xuanxu yes, I have some time to do this. And I like the potential pun on 'fit'. :)
