### 2019-09-29 #1771 @xuanxu
@wtgee Yeah, this is the response from @JonasKemmer to that same question:

>It seems that the only contributor to the software is JonasKemmer.

>>This is not the case, @StephanStock is the main contributer to the code, he just asked me to maintain the code, which is why all commits are from me. Most of the code was written before I joined the project. I just converted it into a publishable form.
