### 2019-09-29 #1771 @wtgee
Hi @xuanxu. Regarding the **Functionality documentation**, I can see that most of the methods have appropriate docstrings but there is nothing like a link out to readthedocs. Is the requirement just regarding docstrings?
