### 2019-09-30 #1771 @xuanxu
@wtgee There should be sufficient documentation to understand the functionality of the software, including a high-level overview in the readme file (see https://joss.readthedocs.io/en/latest/review_criteria.html#documentation). If you feel the included documentation is not enough you can always suggest improvements to the author, I think a link to readthedocs or similar is always convenient.
