### 2019-10-01 547d144
Improved plotsize, fontsize and linewidth
