+++ aliasfinder/af_plots.py
@@ -132,7 +132,7 @@ def get_theta(phase):
 
 
 def plot_phase_clock(theta_obs, theta_sim, range_sims, freq,
-                     main_axis, width=0.3):
+                     main_axis, width=0.23):
     """ Plot function for an inset phase-clock """
     trans = blended_transform_factory(main_axis.transData,
                                       main_axis.transAxes)
