### 2019-10-01 56e4246
implemented scipy's find_peaks instead of detect_peaks
