+++ contributing.md
@@ -1,0 +1,35 @@
+# How to contribute to the AliasFinder
+
+We welcome contributions in the form of bug reports,
+bug fixes, improvements to the documentation, ideas for enhancements, or the enhancements themselves!  
+
+## Did you find a bug?
+
+Please first have a look at [project's GitHub repository issues](https://github.com/JonasKemmer/AliasFinder/issues). 
+If you do not find your issues there please fill in a bug report. 
+Github allows you to classify your issues so that we know if it is a bug report, 
+feature request or feedback to the authors.
+
+* Please include a short Python snippet or explanation for reproducing the problem. If possible attach 
+  the input files that were used to reproduce the error.
+ 
+* Explain the behavior you expected, and how it differs from the real behavior.
+  
+* If applicable, include the error message that was printed on the screen.
+
+
+## Do you request an enhacement or want to give us feedback?
+
+Please follow the same steps as described above for filling in a bug report.
+
+## Do you want to fix the the bug yourself or enhance the code or documentation?
+
+If you wish to contribute to the code, you should submit a pull request. Please test the code beforehand and describe 
+your changes in detail in the pull request. 
+You can find informations of how to fork a repository, clone it and work on the
+changes in the [GitHub guides](https://guides.github.com/activities/forking/).
+
+
+## Meta
+
+Thank you for using AliasFinder and thank you for your will to contribute!
