### 2019-10-01 a42b629
deleted detect_peaks.py since its no longer needed, updated README.md accordingly
