### 2019-10-26 #1771 @Juanlu001
Sorry for the long delay and thanks @StephanStock for your patience, I finished the review. Some comments:

* It would be nice to add a git tag to the repository, to clearly signal what's `v1.0.0`.
* Regarding "Contribution and authorship", I ticket off the box [because of this comment](https://github.com/openjournals/joss-reviews/issues/1724#issuecomment-530867688):

> @StephanStock is the main contributer to the code, he just asked me to maintain the code, which is why all commits are from me. Most of the code was written before I joined the project. I just converted it into a publishable form.

* Regarding "Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?" -- The software is meant to be used as a command line tool and the configuration options are properly described in the example file, so I ticket this item off too.
* The repository contains no automated tests.
  - Perhaps the authors could consider reusing the `Example/` data to create an automatic test using https://github.com/matplotlib/pytest-mpl (to check that the result image matches https://github.com/JonasKemmer/AliasFinder/blob/master/Example/GJ436_2.64d_1.00297d_alias_test.png.
  - To make this automatic (for example, to be checked in every commit), it would require some infrastructure, for example https://travis-ci.org/.
* There is no comparison with other similar software, which suggests that there is probably no other commonly used packages. If that's the case, having it stated somewhere could be perhaps beneficial. "To our knowledge, there are no..."
* The script uses several [Python dependencies](https://github.com/JonasKemmer/AliasFinder#requirements), it would be nice to add corresponding citations to this paper (at least I know that astropy, numpy, matplotlib and SciPy have papers or references that can be cited)

Other than that, the script worked flawlessly, and when instructed to do so it leveraged all 4 cores in my machine. Nice work @StephanStock @JonasKemmer!
