### 2019-11-14 #1771 @xuanxu
@StephanStock / @JonasKemmer: Did you have the opportunity to address the [feedback from @Juanlu001](https://github.com/openjournals/joss-reviews/issues/1771#issuecomment-546592112)?
