### 2019-11-21 83ea5b0
Included changes to the JOSS manuscript.
