### 2019-11-21 9f245e8
corrected JOSS paper references
