### 2019-11-21 #1771 @StephanStock
@Juanlu001 thank you for the review and positive feedback.

@xuanxu we have taken the reviews into account and marked changes in bold in the new paper.md file.

1. We will include a git tag when the paper is accepted and define version 1.0.0
as the version which was accepted.

2. Regarding the request for automated testing, we think that this is not feasible for our script
for the moment. The provided example can be used as a manual test. We will include in the README, how the data can be downloaded when pip install is used.

3. We included the following sentence as suggested by @Juanlu001:

"To our knowledge, there are no other public packages to perform alias testing based on this methodology."

Thanks to the referee for this suggestion!

4. We have added all the appropriate citations to for the packages which we use.

Thanks to @xuanxu  for keeping track of this. If there is anything else we should do, please let us know.
