+++ aliasfinder/af_main_save.py
@@ -1,313 +1,0 @@
-import os
-import sys
-import warnings
-from multiprocessing import Pool
-
-import matplotlib
-import matplotlib.gridspec as gridspec
-import matplotlib.pyplot as plt
-import numpy as np
-import yaml
-from matplotlib.widgets import Cursor
-from matplotlib.ticker import MaxNLocator
-from tqdm import tqdm
-from astropy.table import Table
-
-import aliasfinder.af_calc as af_calc
-import aliasfinder.af_plots as af_plots
-import aliasfinder.af_utils as af_utils
-
-
-__author__ = "Jonas Kemmer, Stephan Stock @ ZAH, Landessternwarte Heidelberg"
-__version__ = "1.0"
-__license__ = "MIT"
-
-
-warnings.filterwarnings("ignore")
-
-#if __name__ == '__main__':
-def main():
-    print(f' \n'
-          f'Welcome to the Alias Finder V{__version__}\n')
-    # Input the parameter set in the parameter file - the list of possible
-    # arguments is defined in the default_params-dictionairy
-    with open(str(sys.argv[1])) as stream:
-        try:
-            input_params = yaml.safe_load(stream)
-        except yaml.YAMLError as exc:
-            print(exc)
-    default_params = {'object_name': '',
-                      'savepath': './',
-                      'rv_files': [],
-                      'test_period': 50,
-                      'sampling_freq': None,
-                      'mc_samples': 100,
-                      'num_processes': 1,
-                      'alias_order': 1,
-                      'panel_width': 1,
-                      'hide_xlabel': True,
-                      'plot_additional_period_axis': True,
-                      'fbeg': 0.0001,
-                      'fend': 2.5,
-                      'power_threshold': 0.05,
-                      'search_phase_range': 0.00005,
-                      'offsets': [0],
-                      'substract': False,
-                      'use_rms_as_jitter': True,
-                      'jitter': 0}
-    params = {}
-    for key in default_params:
-        try:
-            params = {key: input_params[key]}
-        except KeyError:
-            params = {key: default_params[key]})
-            print(f'Parameter {key} is missing in the input file.\n'
-                  f'Value set to default: {key} = {default_params[key]}\n')
-    # Read in the observed RV data
-    times, rvs, rvs_err = af_utils.read_rvs(rv_files, offsets)
-    if sampling_freq is None or sampling_freq == 'None':
-        # First analyze the spectral window function of the data and select a
-        # sampling frequency who's aliases should be analyzed
-        window_freqs = np.linspace(fbeg, fend, len(times)*25)
-        window_powers = af_utils.get_spectral_window_function(window_freqs,
-                                                              times)
-
-        print('Please select the sampling frequency you want to test '
-              'from the plot')
-        fig = plt.figure()
-        ax = fig.add_subplot(1, 1, 1)
-        plt.title('Spectral window function of the data\n'
-                  'Please select the sampling frequency you want to test '
-                  '(Press any button then click)')
-        plt.xlabel('Frequency f [1/d]')
-        plt.ylabel(r'Power W($\nu$)')
-        main_line = ax.plot(window_freqs, window_powers, color='black')
-        cursor = Cursor(ax, useblit=True, color='k', linewidth=1)
-        zoom_ok = False
-        print('\nZoom or pan to view, \npress any button when ready'
-              ' to select the sampling frequency - then click:\n')
-        while not zoom_ok:
-            zoom_ok = plt.waitforbuttonpress()
-        sampling_freq = plt.ginput(1)
-        sampling_freq = np.round(sampling_freq[0][0], 5)
-        plt.axis('auto')
-        plt.autoscale()
-        plt.xlim(0, 1.2)
[…]
