### 2019-11-25 dc67f85
Added instructions for manual testing of the installation.
