### 2019-12-02 #1771 @xuanxu
@StephanStock
>     1. We will include a git tag when the paper is accepted and define version 1.0.0
>        as the version which was accepted.

That's OK. I'll ask you for it when needed.

>     2. Regarding the request for automated testing, we think that this is not feasible for our script
>        for the moment. The provided example can be used as a manual test. We will include in the README, how the data can be downloaded when pip install is used.

OK, if automated tests are not in place, some form of manual procedure to test the functionality of the software should be included. Please document that (in the README is OK, in a separate file linked from the README is also OK).

> we are a bit worried that wtgee did not respond to any message since September 30th.

I've contacted him via email, I hope he'll provide an update soon.
