### 2019-12-03 #1771 @JonasKemmer
Hello @xuanxu and @Juanlu001,
thanks for your comments. Instructions for manual testing are now included in the README.
