### 2020-01-10 8326ad6
JOSS paper fix recommendations
A few grammatical and spelling mistakes in the JOSS paper.
