### 2020-01-10 #1771 @wtgee
Hi folks,

My sincere apologies for the ridiculous delay. Also thanks to @Juanlu001 for earlier reviews. The repository looks much improved from when I had first reviewed it before.  I agree that this is ready, just a few comments:

* Obviously I would like to see automated testing and encourage the authors to get in the future habit of developing tests as you develop software.
* There is a `contributing.md` but no reference on the main landing page about how to contribute or help, etc. Maybe just a small section at the bottom that links to the contributing doc and has some standard "If you want to help..." text.
* There are a number of spelling and grammatical mistakes in the JOSS paper still. I know I'm not supposed to do direct PRs, but it was just as easy to edit the document as to mark where the errors were: https://github.com/JonasKemmer/AliasFinder/pull/8

Other:
* The final plot now includes the Period as the secondary x-axis, which is helpful. Should update the plot on the github page to reflect that.
* The final plot might also be more informative with the residuals plotted. I'm not entirely sure what that would look like (maybe a 4th column with all the residuals from the row?) and don't think you need it for this version, but just a suggestion.

Other than that, things look great! Apologies again for the delay!
