### 2020-01-13 5c8c75d
Added information on how to contribute in README.md
