### 2020-01-13 e8a4ad9
Corrected spelling errors and grammar in paper.md
