### 2020-01-13 #1771 @JonasKemmer
Hey @wtgee,
also thanks to you for your comments. Automated testing is for sure on our bucket list, however at this stage not feasible to implement. We added a reference to the `contributing.md` in the `README.md` as suggested and also included some of your spelling and grammar corrections, where we found it appropriate. Also we updated the example plot. Regarding your comment on the residual plot: we are currently investigating the possibility to include a metric to quantify the differences between data and simulation. For that we will sure make use of the residuals. All the best, Jonas and Stephan
