### 2020-01-13 #1771 @wtgee
Great, thanks @JonasKemmer for the updates.  I realize automated testing wasn't going to make it in, just wanted to encourage you for the future.

@xuanxu I've checked off all my items. I see that @Juanlu001 still has a few checkboxes to fill but I think they had already given the signoff.  Everything looks good to me know as well.

Sorry again for the delay!
