### 2020-01-15 #1771 @StephanStock
Thank you @wtgee and @Juanlu001 for your feedback. It significantly helped to improve this software.
