### 2020-01-15 #1771 @StephanStock
There are still some textpassages in the paper in boldface, can we now make them normal?
