### 2020-01-15 #1771 @xuanxu
> There are still some textpassages in the paper in boldface, can we now make them normal?

@StephanStock Yes, go ahead and please tell whedon to generate the pdf when ready
