### 2020-01-15 12edaba
removed bold-font passages from paper.md
