### 2020-01-15 #1771 @JonasKemmer
Hey @xuanxu, we uploaded the release v1.0 to Zenodo, you can find it here: http://doi.org/10.5281/zenodo.3608898
