### 2020-01-16 #1771 @kyleniemeyer
Hi @StephanStock, I'm doing some final checks on the article before publication.

In the references, it looks like the journal/publication venue is missing for a few: both Astropy articles, Dawson, Trifonov, and Zechmeister. Can you check on that?
