### 2020-01-16 #1771 @kyleniemeyer
Some other minor fixes:
- please add commas after "e.g." -> `e.g.,`
- In the second paragraph, change the Trifonov in-text reference to use the `@author:2001` formatting to remove the parentheses—it will appear like "Trifonov et al. (2018)"
