### 2020-01-16 7ce25c0
Small corrections to paper.md & journal/article venues corrected in .bib
