+++ JOSS_Paper/paper_alias.bib
@@ -1,7 +1,7 @@
 @ARTICLE{Zechmeister2009,
    author = {{Zechmeister}, M. and {K{\"u}rster}, M.},
     title = "{The generalised Lomb-Scargle periodogram. A new formalism for the floating-mean and Keplerian periodograms}",
-  journal = {\aap},
+  journal = {Astronomy {\&} Astrophysics},
 archivePrefix = "arXiv",
    eprint = {0901.2573},
  primaryClass = "astro-ph.IM",
@@ -18,7 +18,7 @@ archivePrefix = "arXiv",
 @ARTICLE{Dawson2010,
    author = {{Dawson}, R.~I. and {Fabrycky}, D.~C.},
     title = "{Radial Velocity Planets De-aliased: A New, Short Period for Super-Earth 55 Cnc e}",
-  journal = {\apj},
+  journal = {Astrophysical Journal},
 archivePrefix = "arXiv",
    eprint = {1005.4050},
  primaryClass = "astro-ph.EP",
@@ -92,7 +92,7 @@ archivePrefix = "arXiv",
          {Wagner}, K. and {Winkler}, J. and {Wolthoff}, V. and {Xu}, W. and
          {Yan}, F. and {Zapatero Osorio}, M.~R.},
         title = "{The CARMENES search for exoplanets around M dwarfs . First visual-channel radial-velocity measurements and orbital parameter updates of seven M-dwarf planetary systems}",
-      journal = {\aap},
+      journal = {Astronomy {\&} Astrophysics},
      keywords = {planetary systems, stars: late-type, stars: low-mass, planets and satellites: dynamical evolution and stability, Astrophysics - Earth and Planetary Astrophysics},
          year = "2018",
         month = "Feb",
@@ -161,7 +161,7 @@ archivePrefix = {arXiv},
          {Weaver}, B.~A. and {Whitmore}, J.~B. and {Woillez}, J. and
          {Zabalza}, V. and {Astropy Contributors}},
         title = "{The Astropy Project: Building an Open-science Project and Status of the v2.0 Core Package}",
-      journal = {\aj},
+      journal = {Astronomical Journal},
      keywords = {methods: data analysis, methods: miscellaneous, methods: statistical, reference systems, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = "2018",
         month = "Sep",
@@ -196,7 +196,7 @@ archivePrefix = {arXiv},
          {Pierfederici}, Francesco and {Pontzen}, Andrew and {Ptak}, Andy and
          {Refsdal}, Brian and {Servillat}, Mathieu and {Streicher}, Ole},
         title = "{Astropy: A community Python package for astronomy}",
-      journal = {\aap},
+      journal = {Astronomy {\&} Astrophysics},
      keywords = {methods: data analysis, methods: miscellaneous, virtual observatory tools, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = "2013",
         month = "Oct",
@@ -271,7 +271,3 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2019arXiv190710121V},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
-
-
-
-
