### 2020-01-16 #1771 @JonasKemmer
@kyleniemeyer nice catch with the references, we updated the files. We also corrected the spelling errors in the paper. Thanks for letting us know.
